Finished function and included coulmnnames
replaced the function that read in data from excel by the one that reads it in form MS Access. Same function name, same return. File is named the same. Old file is saved as well, the function got the suffix "old".
added comments, should be readable

install pyodbc must be added to dependencies.